Force refresh of Decor View to work around Android bug
In Android 14 and newer releases of Android 13 (somewhere between r31 and r38),
a bug was introduced into Android which causes the options menu to open only
on the first call of Activity.openOptionsMenu() and to ignore subsequent
attempts to open the menu again.

For more details, see https://issuetracker.google.com/issues/315761686.

To work around the issue, we set refreshDecorView to true on closing the menu,
in order to force a more complete re-initialisation the next time the menu is
being opened.